Move canonical(Decode|Encode)Pretty from test module
It is used further up the stack in non-test contexts.
